+++ src/UI/Atoms/Small/Small.interface.ts
@@ -1,0 +1,44 @@
+import classNames from 'classnames';
+import { ColorContrastClass } from '../../Particles/Particles';
+import {
+  AddClassDictionary,
+  ColorValues,
+  ContrastValues,
+  TextElements,
+} from '../../types';
+
+//  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
+export interface iSmall {
+  contrast?: ContrastValues;
+  color?: ColorValues;
+  tag?: TextElements;
+  addClass?: AddClassDictionary;
+}
+
+//  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
+export interface Small {
+  contrast: ContrastValues;
+  color: ColorValues;
+  tag: TextElements;
+  addClass: AddClassDictionary;
+}
+
+//  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
+export const dSmall: Small = {
+  contrast: 'highest',
+  color: 'grey',
+  tag: 'span',
+  addClass: {},
+};
+
+//  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
+// Component Classes
+export function cSmall(p: Small): AddClassDictionary {
+  return {
+    small: classNames(
+      'a-Small',
+      ColorContrastClass(p.color, p.contrast, 'txt'),
+      p.addClass.small
+    ),
+  };
+}