+++ src/UI/Atoms/Title/Title.ts
@@ -1,48 +1,25 @@
 /**
- * @module Components/Atoms/Title
+ * @module Atoms/Title
  */
-import { iAnaConfiguration } from '../../../Ana/Ana.interface'
-import { RenderDictionary, ContrastValues, ColorValues } from '../../../types'
-import classNames from 'classnames'
-import { ColorContrastClass } from '../../Particles/Particles'
+import { applyDefaultParameters, StaticChild } from 'ana.js';
+import { Title, cTitle, dTitle, iTitle } from './Title.interface';
+import { a } from '../../ana';
+import './Title.scss';
 
 //  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
-//   _____ _ _   _
-//  |_   _(_) |_| | ___
-//    | | | | __| |/ _ \
-//    | | | | |_| |  __/
-//    |_| |_|\__|_|\___|
-//
+//   ____                                       _
+//  |  _ \ __ _ _ __ __ _  __ _ _ __ __ _ _ __ | |__
+//  | |_) / _` | '__/ _` |/ _` | '__/ _` | '_ \| '_ \
+//  |  __/ (_| | | | (_| | (_| | | | (_| | |_) | | | |
+//  |_|   \__,_|_|  \__,_|\__, |_|  \__,_| .__/|_| |_|
+//                        |___/          |_|
 ////////////////////////////////////////////////////////////////////////////////////////////////////////
-/**
- * A Title is the largest heading of the three heading types, the other two being the Heading and Subheading. It is conventionally used for page title and only one can be found, usually at the top of the page.
- */
-export function rTitle(
-  a: RenderDictionary,
-  config: iAnaConfiguration
-): Function {
-  config
-  return (...children: [Node | string | Function]): Function => {
-    return (param: iTitle = {}): HTMLElement => {
-      // Default values
-      param = {
-        ...{ contrast: 'highest', color: 'grey' },
-        ...param,
-      }
-      let classes = {
-        Title: classNames(
-          'a-Title',
-          ColorContrastClass(param.color, param.contrast, 'txt')
-        ).split(' '),
-      }
 
-      return a.div(...classes.Title)(...children)
-    }
-  }
-}
+export const rTitle =
+  (param: iTitle = {}): Function =>
+  (...children: StaticChild[]): HTMLElement => {
+    let p: Title = applyDefaultParameters<Title, iTitle>(dTitle, param);
+    let c = cTitle(p);
 
-//  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -  -
-export interface iTitle {
-  contrast?: ContrastValues
-  color?: ColorValues
-}
+    return a[p.tag](c.title)(...children);
+  };