+++ src/main.ts
@@ -1,10 +1,18 @@
-import { GenericData} from 'ana.js'
-import { app } from "./UI/ana";
+import { GenericData } from 'ana.js';
+import { app } from './UI/ana';
 import { components, Components } from './UI/Components';
 
-app.components = { ...components }
+app.components = { ...components };
 const a = app.render<Components>();
 
-app.init({},
-  (d: GenericData) => a.Box(a.span()('Test'))()
+app.init({}, (d: GenericData) =>
+  a.Box()(
+    a.Display()('Display'),
+    a.Title()('Title'),
+    a.Heading()('Heading'),
+    a.Subheading()('Subheading'),
+    a.Paragraph()('Paragraph'),
+    a.Label()('Label'),
+    a.Small()('Small')
+  )
 );